+++ @yundeving/story-book-demo-ui/ui/components/index.ts
@@ -26,6 +26,5 @@ export {
   SelectValue,
 } from "./Select";
 export { default as Skeleton } from "./Skeleton";
-export { default as StatusButton } from "./StatusButton";
 export { default as Tooltip } from "./Tooltip";
 export { default as Link } from "./Link";